Directory resources to types
