+++ workos/resources/webhooks.py
@@ -1,7 +1,7 @@
 from typing import Generic, Literal, Union
 from pydantic import Field
 from typing_extensions import Annotated
-from workos.resources.directory_sync import DirectoryGroup
+from workos.types.directory_sync import DirectoryGroup
 from workos.resources.events import EventPayload
 from workos.resources.user_management import OrganizationMembership, User
 from workos.resources.workos_model import WorkOSModel
