+++ workos/types/directory_sync/directory.py
@@ -1,0 +1,20 @@
+from typing import Literal, Optional
+from workos.resources.workos_model import WorkOSModel
+from workos.types.directory_sync.directory_state import DirectoryState
+from workos.types.directory_sync.directory_type import DirectoryType
+from workos.typing.literals import LiteralOrUntyped
+
+
+class Directory(WorkOSModel):
+    """Representation of a Directory Response as returned by WorkOS through the Directory Sync feature."""
+
+    id: str
+    object: Literal["directory"]
+    domain: Optional[str] = None
+    name: str
+    organization_id: str
+    external_key: str
+    state: LiteralOrUntyped[DirectoryState]
+    type: LiteralOrUntyped[DirectoryType]
+    created_at: str
+    updated_at: str
