+++ workos/types/directory_sync/directory_type.py
@@ -1,0 +1,24 @@
+from typing import Literal
+
+
+DirectoryType = Literal[
+    "azure scim v2.0",
+    "bamboohr",
+    "breathe hr",
+    "cezanne hr",
+    "cyperark scim v2.0",
+    "fourth hr",
+    "generic scim v2.0",
+    "gsuite directory",
+    "hibob",
+    "jump cloud scim v2.0",
+    "okta scim v2.0",
+    "onelogin scim v2.0",
+    "people hr",
+    "personio",
+    "pingfederate scim v2.0",
+    "rippling v2.0",
+    "sftp",
+    "sftp workday",
+    "workday",
+]
