+++ workos/types/events/directory_group_with_previous_attributes.py
@@ -1,4 +1,4 @@
-from workos.resources.directory_sync import DirectoryGroup
+from workos.types.directory_sync import DirectoryGroup
 from workos.types.events.previous_attributes import PreviousAttributes
