+++ workos/types/events/directory_payload.py
@@ -1,5 +1,5 @@
 from typing import Literal
-from workos.resources.directory_sync import DirectoryType
+from workos.types.directory_sync import DirectoryType
 from workos.resources.workos_model import WorkOSModel
 from workos.types.directory_sync.directory_state import DirectoryState
 from workos.typing.literals import LiteralOrUntyped
